// rw-app/src/utilities/interfaces/IDavisVantagePro2Plus.ts
import { ISensorData } from "./ISensorData";

export interface IDavisVantagePro2Plus extends ISensorData {
    rx_state: number;
    wind_speed_hi_last_2_min: number;
    hum: number;
    wind_dir_at_hi_speed_last_10_min: number;
    wind_chill: number;
    rain_rate_hi_last_15_min_clicks: number;
    thw_index: number;
    wind_dir_scalar_avg_last_10_min: number;
    rain_size: number;
    uv_index: number;
    wind_speed_last: number;
    rainfall_last_60_min_clicks: number;
    wet_bulb: number;
    rainfall_monthly_clicks: number;
    wind_speed_avg_last_10_min: number;
    wind_dir_at_hi_speed_last_2_min: number;
    rainfall_daily_in: number;
    wind_dir_last: number;
    rainfall_daily_mm: number;
    rain_storm_last_clicks: number;
    tx_id: number;
    rain_storm_last_start_at: number;
    RainStormLastStartAt: Date;
    rain_rate_hi_clicks: number;
    rainfall_last_15_min_in: number;
    rainfall_daily_clicks: number;
    dew_point: number;
    rainfall_last_15_min_mm: number;
    rain_rate_hi_in: number;
    rain_storm_clicks: number;
    rain_rate_hi_mm: number;
    rainfall_year_clicks: number;
    rain_storm_in: number;
    rain_storm_last_end_at: number;
    RainStormLastEndAt: Date;
    rain_storm_mm: number;
    wind_dir_scalar_avg_last_2_min: number;
    heat_index: number;
    rainfall_last_24_hr_in: number;
    rainfall_last_60_min_mm: number;
    trans_battery_flag: number;
    rainfall_last_60_min_in: number;
    rain_storm_start_time: number;
    RainStormStartTime: Date;
    rainfall_last_24_hr_mm: number;
    rainfall_year_in: number;
    wind_speed_hi_last_10_min: number;
    rainfall_last_15_min_clicks: number;
    rainfall_year_mm: number;
    wind_dir_scalar_avg_last_1_min: number;
    temp: number;
    wind_speed_avg_last_2_min: number;
    solar_rad: number;
    rainfall_monthly_mm: number;
    rain_storm_last_mm: number;
    wind_speed_avg_last_1_min: number;
    thsw_index: number;
    rainfall_monthly_in: number;
    rain_rate_last_mm: number;
    rain_rate_last_clicks: number;
    rainfall_last_24_hr_clicks: number;
    rain_storm_last_in: number;
    rain_rate_last_in: number;
    rain_rate_hi_last_15_min_mm: number;
    rain_rate_hi_last_15_min_in: number;
    ts: number;
    TimeStamp: Date;
    Type: number;
    Description: string;
}